+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: cmariot <cmariot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/16 14:08:16 by cmariot           #+#    #+#             */
-/*   Updated: 2022/01/04 08:50:54 by cmariot          ###   ########.fr       */
+/*   Updated: 2022/01/04 16:32:15 by cmariot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
